int64 indexing: partial fixes for ArraySlice, random, incrementor
- ArraySlice.cs: Change Allocate count parameter handling for long
- UnmanagedMemoryBlock: Adjust for long count
- np.random.choice.cs: Add explicit casts for int64 indices
- np.random.shuffle.cs: Update index handling for long
- ValueCoordinatesIncrementor.cs: Add long[] Index property
- NDArray.cs: Remove duplicate/dead code (112 lines)

+++ src/NumSharp.Core/Backends/Unmanaged/ArraySlice.cs
@@ -480,7 +480,7 @@ namespace NumSharp.Backends.Unmanaged
         /// <param name="count">How many items this array will have (aka Count).</param>
         /// <param name="fill">The item to fill the newly allocated memory with.</param>
         /// <returns>A newly allocated array.</returns>
-        public static ArraySlice<T> Allocate<T>(int count, T fill) where T : unmanaged
+        public static ArraySlice<T> Allocate<T>(long count, T fill) where T : unmanaged
             => new ArraySlice<T>(new UnmanagedMemoryBlock<T>(count, fill));
 
         /// <summary>
@@ -497,7 +497,7 @@ namespace NumSharp.Backends.Unmanaged
         /// </summary>
         /// <param name="count">How many items this array will have (aka Count).</param>
         /// <returns>A newly allocated array.</returns>
-        public static ArraySlice<T> Allocate<T>(int count) where T : unmanaged
+        public static ArraySlice<T> Allocate<T>(long count) where T : unmanaged
             => new ArraySlice<T>(new UnmanagedMemoryBlock<T>(count));
 
         /// <summary>
@@ -506,7 +506,7 @@ namespace NumSharp.Backends.Unmanaged
         /// <param name="address">The address at which the memory block starts</param>
         /// <param name="count">The count of items of type <typeparamref name="T"/> (not bytes count)</param>
         /// <returns>A wrapped memory block as <see cref="ArraySlice{T}"/></returns>
-        public static unsafe ArraySlice<T> Wrap<T>(T* address, int count) where T : unmanaged => new ArraySlice<T>(new UnmanagedMemoryBlock<T>(address, count));
+        public static unsafe ArraySlice<T> Wrap<T>(T* address, long count) where T : unmanaged => new ArraySlice<T>(new UnmanagedMemoryBlock<T>(address, count));
 
         /// <summary>
         ///     Wrap around a <paramref name="address"/> with given <paramref name="count"/> without claiming ownership of the address.
@@ -514,6 +514,6 @@ namespace NumSharp.Backends.Unmanaged
         /// <param name="address">The address at which the memory block starts</param>
         /// <param name="count">The count of items of type <typeparamref name="T"/> (not bytes count)</param>
         /// <returns>A wrapped memory block as <see cref="ArraySlice{T}"/></returns>
-        public static unsafe ArraySlice<T> Wrap<T>(void* address, int count) where T : unmanaged => new ArraySlice<T>(new UnmanagedMemoryBlock<T>((T*)address, count));
+        public static unsafe ArraySlice<T> Wrap<T>(void* address, long count) where T : unmanaged => new ArraySlice<T>(new UnmanagedMemoryBlock<T>((T*)address, count));
     }
 }
